+++ app/src/main/java/org/simple/clinic/drugstockreminders/DrugStockReminderApiModule.kt
@@ -5,10 +5,13 @@ import com.f2prateek.rx.preferences2.RxSharedPreferences
 import dagger.Module
 import dagger.Provides
 import org.simple.clinic.main.TypedPreference
-import org.simple.clinic.main.TypedPreference.Type.UpdateDrugStockReportsMonth
-import org.simple.clinic.util.preference.StringPreferenceConverter
+import org.simple.clinic.main.TypedPreference.Type.DrugStockReportLastCheckedAt
+import org.simple.clinic.main.TypedPreference.Type.IsDrugStockReportFilled
+import org.simple.clinic.util.preference.BooleanPreferenceConverter
+import org.simple.clinic.util.preference.InstantRxPreferencesConverter
 import org.simple.clinic.util.preference.getOptional
 import retrofit2.Retrofit
+import java.time.Instant
 import java.util.Optional
 import javax.inject.Named
 
@@ -20,11 +23,15 @@ object DrugStockReminderApiModule {
     return retrofit.create(DrugStockReminderApi::class.java)
   }
 
-  @TypedPreference(UpdateDrugStockReportsMonth)
+  @TypedPreference(DrugStockReportLastCheckedAt)
   @Provides
-  fun updateDrugStockReportsMonth(
-      rxSharedPreferences: RxSharedPreferences
-  ): Preference<Optional<String>> {
-    return rxSharedPreferences.getOptional("drug_stock_report_month", StringPreferenceConverter())
+  fun drugStockReportLastCheckAtPreference(rxSharedPreferences: RxSharedPreferences): Preference<Instant> {
+    return rxSharedPreferences.getObject("drug_stock_report_last_checked_at", Instant.EPOCH, InstantRxPreferencesConverter())
+  }
+
+  @TypedPreference(IsDrugStockReportFilled)
+  @Provides
+  fun isDrugStockReportFilledPreference(rxSharedPreferences: RxSharedPreferences): Preference<Optional<Boolean>> {
+    return rxSharedPreferences.getOptional("is_drug_stock_report_filled", BooleanPreferenceConverter(), Optional.empty())
   }
 }